solutions/0010.regular-expression-matching/solution.go
package pb0010

func isMatch(s string, p string) bool {
	n, m := len(s), len(p)
	dp := make([][]bool, n+1)
	for i := range dp {
		dp[i] = make([]bool, m+1)
	}
	dp[0][0] = true
	for i := 0; i < n+1; i++ {
		for j := 1; j < m+1; j++ {
			if i == 0 {
				if p[j-1] == '*' {
					dp[i][j] = dp[i][j-2]
				}
				continue
			}
			if s[i-1] == p[j-1] || p[j-1] == '.' {
				dp[i][j] = dp[i-1][j-1]
			} else if p[j-1] == '*' {
				dp[i][j] = dp[i][j-2]
				if s[i-1] == p[j-2] || p[j-2] == '.' {
					dp[i][j] = dp[i][j] || dp[i-1][j]
				}
			}
		}
	}
	return dp[n][m]
}
